Implement authorization policies and ownership validation across controllers

+++ Wasfaty.API/Controllers/DispenseRecordController.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 using Wasfaty.Application.Constants;
 using Wasfaty.Application.DTOs.DispenseRecords;
 using Wasfaty.Application.DTOs.Prescriptions;
@@ -8,27 +9,34 @@ using Wasfaty.Application.Interfaces.IServices;
 [Route("api/[controller]")]
 [ApiController]
 //[Authorize(Roles = Roles.Admin)]
-[Authorize(Roles = Roles.Admin + "," + Roles.Pharmacist)] // استثناء
-
+//[Authorize(Roles = Roles.Admin + "," + Roles.Pharmacist)] // استثناء
+[Authorize]
 public class DispenseRecordController : ControllerBase
 {
     private readonly IDispenseRecordService _dispenseRecordService;
     private readonly IPrescriptionService _prescriptionService;
     private readonly IPharmacistService _pharmacistService;
     private readonly IPharmacyService _pharmacyService;
+    private readonly IAuthorizationService _authorizationService;
 
     public DispenseRecordController(IDispenseRecordService dispenseRecordService,
         IPrescriptionService prescriptionService,
         IPharmacistService pharmacistService,
-        IPharmacyService pharmacyService)
+        IPharmacyService pharmacyService,
+        IAuthorizationService authorizationService)
     {
         _dispenseRecordService = dispenseRecordService;
         _prescriptionService = prescriptionService;
         _pharmacistService = pharmacistService;
         _pharmacyService = pharmacyService;
+        _authorizationService = authorizationService;
     }
 
+    private int GetUserId() =>
+      int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)!.Value);
+
     // GET: api/dispenserecords
+    [Authorize(Policy = "AdminRole")]
     [HttpGet("All", Name = "GetAllDispenseRecor")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
@@ -57,22 +65,60 @@ public class DispenseRecordController : ControllerBase
         }
 
         var dispenseRecord = await _dispenseRecordService.GetByIdAsync(id);
+
         if (dispenseRecord == null)
-        {
-            return NotFound($"DispenseRecord with ID {id} not found.");
-        }
+            return NotFound();
+
+        var auth = await _authorizationService.AuthorizeAsync(
+            User, dispenseRecord, "CanAccessDispenseRecord");
+
+        if (!auth.Succeeded)
+            return Forbid();
+
         return Ok(dispenseRecord);
     }
 
     // POST: api/dispenserecords
+    [Authorize(Policy = "AdminOrPharmacistRole")]
     [HttpPost("CreateDispenseRecord")]
     [ProducesResponseType(StatusCodes.Status201Created)]
     [ProducesResponseType(StatusCodes.Status400BadRequest)]
     public async Task<ActionResult<DispenseRecordDto>> CreateDispenseRecord([FromBody] CreateDispenseRecordDto dispenseRecordDto)
     {
+
         if (dispenseRecordDto == null)
         {
             return BadRequest("Invalid DispenseRecord data.");
+        }
+
+        var userId = GetUserId();
+
+        var role = User.FindFirst(ClaimTypes.Role)?.Value;
+
+        if (role == "Admin")
+        {
+            var pharmacist = await _pharmacistService.GetByIdAsync(dispenseRecordDto.PharmacistId);
+
+            if (pharmacist == null)
+            {
+                return BadRequest("الصيدلي مش موجود");
+
+            }
+
+            dispenseRecordDto.PharmacistId = pharmacist.Id;
+
+        }
+
+        if (role == "Pharmacist")
+        {
+            var pharmacist = await _pharmacistService
+                .GetPharmacistByUserIdAsync(userId);
+
+            if (pharmacist == null)
+                return Unauthorized();
+
+            dispenseRecordDto.PharmacistId = pharmacist.Id;
+
         }
 
         var prescription = await _prescriptionService.GetByIdAsync(dispenseRecordDto.PrescriptionId);
@@ -83,13 +129,11 @@ public class DispenseRecordController : ControllerBase
 
         }
 
-        var pharmacist = await _pharmacistService.GetByIdAsync(dispenseRecordDto.PharmacistId);
 
-        if (pharmacist == null)
-        {
-            return BadRequest("الصيدلي مش موجود");
+        if (prescription.IsDispensed)
+            return BadRequest("تم صرف الوصفة مسبقاً");
 
-        }
+
 
         var pharmacy = await _pharmacyService.GetByIdAsync(dispenseRecordDto.PharmacyId);
 
@@ -98,6 +142,8 @@ public class DispenseRecordController : ControllerBase
             return BadRequest("الصيدلية مش موجودة");
 
         }
+
+        dispenseRecordDto.PharmacyId = pharmacy.Id;
 
         var createdDispenseRecord = await _dispenseRecordService.CreateAsync(dispenseRecordDto);
 
@@ -110,6 +156,7 @@ public class DispenseRecordController : ControllerBase
     }
 
     // PUT: api/dispenserecords/{id}
+    [Authorize(Policy = "AdminOrPharmacistRole")]
     [HttpPut("{id}", Name = "UpdateDispenseRecord")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
@@ -152,6 +199,12 @@ public class DispenseRecordController : ControllerBase
             return NotFound($"DispenseRecord with ID {id} not found.");
         }
 
+        var auth = await _authorizationService.AuthorizeAsync(
+          User, existingDispenseRecord, "CanEditDispenseRecord");
+
+        if (!auth.Succeeded)
+            return Forbid();
+
         var updatedDispenseRecord = await _dispenseRecordService.UpdateAsync(id, dispenseRecordDto);
         if (updatedDispenseRecord == null)
         {
@@ -162,6 +215,7 @@ public class DispenseRecordController : ControllerBase
     }
 
     // DELETE: api/dispenserecords/{id}
+    [Authorize(Policy = "AdminRole")]
     [HttpDelete("{id}", Name = "DeleteDispenseRecord")]
     [ProducesResponseType(StatusCodes.Status204NoContent)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
@@ -179,6 +233,12 @@ public class DispenseRecordController : ControllerBase
             {
                 return NotFound($"DispenseRecord center with ID {id} not found.");
             }
+
+            var auth = await _authorizationService.AuthorizeAsync(
+     User, existingDispenseRecord, "CanEditDispenseRecord");
+
+            if (!auth.Succeeded)
+                return Forbid();
 
             if (await _dispenseRecordService.DeleteAsync(id))
             {
